% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMsigma.R
\name{ADMMsigma}
\alias{ADMMsigma}
\title{ADMM penalized precision matrix estimation (using ADMMsigmac)}
\usage{
ADMMsigma(X = NULL, S = NULL, lam = 10^seq(-5, 5, 0.5), alpha = seq(0,
  1, 0.1), diagonal = FALSE, rho = 2, mu = 10, tau1 = 2, tau2 = 2,
  crit = "ADMM", tol1 = 1e-04, tol2 = 1e-04, maxit = 1000, K = 5,
  cores = 1, quiet = TRUE)
}
\arguments{
\item{X}{data matrix}

\item{S}{option to specify sample covariance matrix (denominator n)}

\item{lam}{tuning parameter for penalty. Defaults to 10^seq(-5, 5, 0.5)}

\item{alpha}{elasticnet mixing parameter [0, 1]: 0 = ridge, 1 = lasso/bridge. Defaults to seq(-1, 1, 0.1)}

\item{diagonal}{option to penalize diagonal elements. Defaults to FALSE}

\item{rho}{initial step size for ADMM}

\item{mu}{factor for primal and residual norms}

\item{tau1}{adjustment for rho}

\item{tau2}{adjustment for rho}

\item{crit}{criterion for convergence c('ADMM', 'grad', 'loglik'). Option crit != 'ADMM' will use tol1 as tolerance. Default is 'ADMM'}

\item{tol1}{absolute tolerance. Defaults to 1e-4}

\item{tol2}{relative tolerance. Defaults to 1e-4}

\item{maxit}{maximum number of iterations}

\item{K}{specify the number of folds for cross validation}

\item{cores}{option to run CV in parallel. Defaults to cores = 1}

\item{quiet}{specify whether the function returns progress of CV or not}
}
\value{
iterations, lam, omega, and gradient
}
\description{
Penalized Gaussian likelihood precision matrix estimation using the ADMM algorithm.
}
\examples{
ADMM_sigma(X, lam = 0.1, rho = 10)
}
