% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ADMMsigmac}
\alias{ADMMsigmac}
\title{ADMM penalized precision matrix estimation (c++)}
\usage{
ADMMsigmac(S, lam, alpha = 1, rho = 2, mu = 10, tau1 = 2, tau2 = 2,
  crit = "ADMM", tol1 = 1e-04, tol2 = 1e-04, maxit = 1000L)
}
\arguments{
\item{S}{option to specify sample covariance matrix (denominator n)}

\item{lam}{tuning parameter for penalty}

\item{alpha}{elasticnet mixing parameter [0, 1]: 0 = ridge, 1 = lasso/bridge}

\item{rho}{initial step size for ADMM}

\item{mu}{factor for primal and residual norms}

\item{tau1}{adjustment for rho}

\item{tau2}{adjustment for rho}

\item{crit}{criterion for convergence c("ADMM", "grad", "lik"). Option crit != "ADMM" will use tol1 as tolerance. Defaults to "ADMM"}

\item{tol1}{absolute tolerance. Defaults to 1e-4}

\item{tol2}{relative tolerance. Defaults to 1e-4}

\item{maxit}{maximum number of iterations}

\item{X}{data matrix}
}
\value{
iterations, lam, omega
}
\description{
Penalized Gaussian likelihood precision matrix estimation using the ADMM algorithm.
}
\examples{
ADMM_sigmac(X, lam = 0.1)

}
