% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMsigma.R
\name{CV_ADMMsigma}
\alias{CV_ADMMsigma}
\title{CV ADMM penalized precision matrix estimation}
\usage{
CV_ADMMsigma(X, lam, alpha = 1, rho = 2, mu = 10, tau1 = 2, tau2 = 2,
  crit = "ADMM", tol1 = 1e-04, tol2 = 1e-04, maxit = 1000, ind = NULL,
  K = 3, quiet = TRUE)
}
\arguments{
\item{X}{matrix or data frame. This is the n x p column matrix where the rows are a realization of n independent copies of a p-variate random vector}

\item{lam}{tuning parameter for penalty. Defaults to 10^seq(-5, 5, 0.5)}

\item{alpha}{elasticnet mixing parameter [0, 1]: 0 = ridge, 1 = lasso/bridge}

\item{rho}{initial step size for ADMM}

\item{mu}{factor for primal and residual norms}

\item{tau1}{adjustment for rho}

\item{tau2}{adjustment for rho}

\item{crit}{criterion for convergence c('ADMM', 'grad', 'lik'). Option crit != 'ADMM' will use tol1 as tolerance. Defaults to 'ADMM'}

\item{tol1}{absolute tolerance. Defaults to 1e-4}

\item{tol2}{relative tolerance. Defaults to 1e-4}

\item{maxit}{maximum number of iterations}

\item{ind}{vector of a permutation of 1,..,n for CV}

\item{K}{specify the number of folds for cross validation}

\item{quiet}{specify whether the function returns progress of CV or not}
}
\value{
iterations, lam, S, Omega, and cv.errors
}
\description{
Cross validation function for ADMM_sigma.
}
\examples{
CV_sigma_ridge(X, lam = seq(0.1, 3, 0.1))

}
