% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CV_RIDGEsigmac}
\alias{CV_RIDGEsigmac}
\title{CV ridge penalized precision matrix estimation (c++)}
\usage{
CV_RIDGEsigmac(X, lam, K = 3L, quiet = TRUE)
}
\arguments{
\item{X}{option to provide a nxp matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{lam}{tuning parameter for ridge penalty. Defaults to grid of values \code{10^seq(-5, 5, 0.5)}.}

\item{K}{specify the number of folds for cross validation.}

\item{quiet}{specify whether the function returns progress of CV or not.}
}
\value{
list of returns includes:
\item{lam}{optimal tuning parameter.}
\item{cv.error}{cross validation error for optimal parameters.}
\item{cv.errors}{cross validation errors.}
}
\description{
Cross validation function for RIDGEsigma.
}
\keyword{internal}
