% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_ridge.R
\name{CV_sigma_ridge}
\alias{CV_sigma_ridge}
\title{CV Ridge-penalized precision matrix estimation}
\usage{
CV_sigma_ridge(X, lam, ind = NULL, K = 5, quiet = TRUE)
}
\arguments{
\item{X}{matrix or data frame. This is the n x p column matrix where the rows are a realization of n independent copies of a p-variate random vector}

\item{lam}{tuning parameters for ridge regularization term.}

\item{ind}{vector of a permutation of 1,..,n}

\item{K}{specify the number of folds for cross validation}

\item{quiet}{specify whether the function returns progress or not}
}
\value{
omega hat matrix, best lambda, CV error, vector of lambdas
}
\description{
Cross validation function for sigma_ridge.
}
\examples{
CV_sigma_ridge(X, lam = seq(0.1, 3, 0.1))

}
