% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CV_sigma_ridgec}
\alias{CV_sigma_ridgec}
\title{CV ridge penalized precision matrix estimation (c++)}
\usage{
CV_sigma_ridgec(X, lam, K = 3L, quiet = TRUE)
}
\arguments{
\item{X}{matrix or data frame. This is the n x p column matrix where the rows are a realization of n independent copies of a p-variate random vector}

\item{lam}{tuning parameter for penalty. Defaults to 10^seq(-5, 5, 0.5)}

\item{K}{specify the number of folds for cross validation}

\item{quiet}{specify whether the function returns progress of CV or not}
}
\value{
iterations, lam, S, Omega, and cv.errors
}
\description{
Cross validation function for RIDGEsigma.
}
\examples{
CV_sigma_ridge(X, lam = seq(0.1, 3, 0.1))

}
