% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVP_ADMMsigmac}
\alias{CVP_ADMMsigmac}
\title{CV (no folds) ADMM penalized precision matrix estimation (c++)}
\usage{
CVP_ADMMsigmac(S_train, S_valid, lam, alpha, rho = 2, mu = 10, tau1 = 2,
  tau2 = 2, crit = "ADMM", tol1 = 1e-04, tol2 = 1e-04, maxit = 1000L,
  K = 5L, quiet = TRUE)
}
\arguments{
\item{S_train}{matrix or data frame. This is pxp sample covariance for training data}

\item{S_valid}{matrix or data frame. This is pxp sample covariance for validation data}

\item{lam}{tuning parameter for penalty. Defaults to 10^seq(-5, 5, 0.5)}

\item{alpha}{elasticnet mixing parameter [0, 1]: 0 = ridge, 1 = lasso/bridge}

\item{rho}{initial step size for ADMM}

\item{mu}{factor for primal and residual norms}

\item{tau1}{adjustment for rho}

\item{tau2}{adjustment for rho}

\item{crit}{criterion for convergence c('ADMM', 'grad', 'lik'). Option crit != 'ADMM' will use tol1 as tolerance. Defaults to 'ADMM'}

\item{tol1}{absolute tolerance. Defaults to 1e-4}

\item{tol2}{relative tolerance. Defaults to 1e-4}

\item{maxit}{maximum number of iterations}

\item{quiet}{specify whether the function returns progress of CV or not}
}
\value{
iterations, lam, S, Omega, and cv.errors
}
\description{
Cross validation (no folds) function for ADMM_sigma. This function is to be used with ParallelCV.
}
