% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVP_RIDGEc}
\alias{CVP_RIDGEc}
\title{CV (no folds) RIDGE penalized precision matrix estimation (c++)}
\usage{
CVP_RIDGEc(n, S_train, S_valid, lam, trace = "none")
}
\arguments{
\item{n}{sample size for X_valid (used to calculate CV_error)}

\item{S_train}{pxp sample covariance matrix for training data (denominator n).}

\item{S_valid}{pxp sample covariance matrix for validation data (denominator n).}

\item{lam}{positive tuning parameters for ridge penalty. If a vector of parameters is provided, they should be in increasing order.}

\item{trace}{option to display progress of CV. Choose one of \code{progress} to print a progress bar, \code{print} to print completed tuning parameters, or \code{none}.}
}
\value{
cross validation errors (negative validation likelihood)
}
\description{
Cross validation (no folds) function for RIDGEsigma. This function is to be used with CVP_RIDGE.
}
\keyword{internal}
