% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{ParallelCV_RIDGE}
\alias{ParallelCV_RIDGE}
\title{Parallel Ridge CV (uses CV_RIDGEsigmac)}
\usage{
ParallelCV_RIDGE(X = NULL, lam = 10^seq(-5, 5, 0.5), K = 5, cores = 1,
  quiet = TRUE)
}
\arguments{
\item{X}{nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{lam}{tuning parameter for elastic net penalty. Defaults to grid of values \code{10^seq(-5, 5, 0.5)}.}

\item{K}{specify the number of folds for cross validation.}

\item{cores}{option to run CV in parallel. Defaults to \code{cores = 1}.}

\item{quiet}{specify whether the function returns progress of CV or not.}
}
\value{
returns list of returns which includes:
\item{lam}{optimal tuning parameter.}
\item{cv.error}{cross validation error for optimal parameters.}
\item{cv.errors}{cross validation errors.}
}
\description{
Parallel implementation of cross validation for RIDGEsigma.
}
\keyword{internal}
