% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMsigma.R
\name{plot.ADMMsigma}
\alias{plot.ADMMsigma}
\title{Plot ADMMsigma object}
\usage{
\method{plot}{ADMMsigma}(x, footnote = TRUE, ...)
}
\arguments{
\item{x}{ADMMsigma class object}

\item{footnote}{option to print footnote of optimal values}
}
\description{
produces a heat plot for the cross validation errors
}
