% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMsigma.R
\name{print.ADMM}
\alias{print.ADMM}
\title{Print ADMM object}
\usage{
\method{print}{ADMM}(x, ...)
}
\arguments{
\item{x}{class object ADMM}

\item{...}{additional arguments.}
}
\description{
Prints ADMM object and suppresses output if needed.
}
\keyword{internal}
