% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMsigma.R
\name{print.ADMMsigma}
\alias{print.ADMMsigma}
\title{Print ADMMsigma object}
\usage{
\method{print}{ADMMsigma}(x, ...)
}
\arguments{
\item{x}{class object ADMMsigma}

\item{...}{additional arguments.}
}
\description{
prints ADMMsigma object and suppresses output if needed.
}
\keyword{internal}
