% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDGEsigma.R
\name{print.RIDGE}
\alias{print.RIDGE}
\title{Print RIDGE object}
\usage{
\method{print}{RIDGE}(x, ...)
}
\arguments{
\item{x}{class object RIDGE.}

\item{...}{additional arguments.}
}
\description{
Prints RIDGE object and suppresses output if needed.
}
\keyword{internal}
