% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDGEsigma.R
\name{print.RIDGEsigma}
\alias{print.RIDGEsigma}
\title{Print RIDGEsigma object}
\usage{
\method{print}{RIDGEsigma}(x, ...)
}
\arguments{
\item{x}{class object RIDGEsigma.}

\item{...}{additional arguments.}
}
\description{
Prints RIDGEsigma object and suppresses output if needed.
}
\keyword{internal}
