% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDGEsigma.R
\name{RIDGEsigma}
\alias{RIDGEsigma}
\title{Ridge penalized precision matrix estimation (using RIDGEsigmac)}
\usage{
RIDGEsigma(X = NULL, S = NULL, lam = 10^seq(-5, 5, 0.5), K = 3,
  quiet = TRUE)
}
\arguments{
\item{X}{data matrix}

\item{S}{option to specify sample covariance matrix (denominator n)}

\item{lam}{tuning parameter for penalty. Defaults to 10^seq(-5, 5, 0.5)}

\item{K}{specify the number of folds for cross validation}

\item{quiet}{specify whether the function returns progress of CV or not}
}
\value{
lam, omega, and gradient
}
\description{
Penalized Gaussian likelihood precision matrix estimation using the ADMM algorithm.
}
\examples{
RIDGEsigma(X, lam = 0.1)
}
