% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RIDGEsigmac}
\alias{RIDGEsigmac}
\title{Ridge-penalized precision matrix estimation (c++)}
\usage{
RIDGEsigmac(S, lam)
}
\arguments{
\item{S}{sample covariance matrix (denominator n)}

\item{lam}{tuning parameter for penalty}
}
\value{
matrix of omega hat
}
\description{
Ridge-penalized Gaussian likelihood precision matrix estimation. Augmented from Adam Rothman's STAT 8931 code.
}
\examples{
n = nrow(X)
RIDGEsigmac(S = (n-1)/n*cov(X), lam = 0.1)

}
