% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_ridge.R
\name{sigma_ridge}
\alias{sigma_ridge}
\title{Ridge-penalized precision matrix estimation}
\usage{
sigma_ridge(S, lam)
}
\arguments{
\item{S}{sample covariance matrix (denominator n)}

\item{lam}{tuning parameter for penalty}
}
\value{
matrix of omega hat
}
\description{
Ridge-penalized Gaussian likelihood precision matrix estimation.
}
\examples{
n = nrow(X)
sigma_ridge(S = (n-1)/n*cov(X), lam = 0.1)
}
