% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softc}
\alias{softc}
\title{Soft threshold (elementwise) (c++)}
\usage{
softc(s, tau)
}
\arguments{
\item{s}{scalar}

\item{tau}{scalar}
}
\value{
scalar
}
\description{
Elementwise soft thresholding function. Augmented from Adam Rothman's STAT 8931 code.
}
\examples{
softc(10, 5)

}
