% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softmatrixc}
\alias{softmatrixc}
\title{Soft threshold (matrix) (c++)}
\usage{
softmatrixc(S, tau)
}
\arguments{
\item{tau}{scalar}

\item{s}{matrix}
}
\value{
soft threshold s matrix
}
\description{
Matrix soft thresholding function. Requires `softc`. Augmented from Adam Rothman's STAT 8931 code.
}
\examples{
softmatrixc(10, 5)

}
