% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVP_RIDGEsigmac}
\alias{CVP_RIDGEsigmac}
\title{CV (no folds) RIDGE penalized precision matrix estimation (c++)}
\usage{
CVP_RIDGEsigmac(S_train, S_valid, lam, quiet = TRUE)
}
\arguments{
\item{S_train}{pxp sample covariance matrix for training data (denominator n).}

\item{S_valid}{pxp sample covariance matrix for validation data (denominator n).}

\item{lam}{tuning parameter for elastic net penalty. Defaults to grid of values \code{10^seq(-5, 5, 0.5)}.}

\item{quiet}{specify whether the function returns progress of CV or not.}
}
\value{
cross validation errors
}
\description{
Cross validation (no folds) function for RIDGEsigma. This function is to be used with ParallelCV_RIDGE.
}
\keyword{internal}
