% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDGEsigma.R
\name{plot.RIDGEsigma}
\alias{plot.RIDGEsigma}
\title{Plot RIDGEsigma object}
\usage{
\method{plot}{RIDGEsigma}(x, ...)
}
\arguments{
\item{x}{RIDGEsigma class object}
}
\description{
produces a heat plot for the cross validation errors
}
